Policy implementation to controllers and few other fixes

+++ app/Http/Controllers/Procurement/PlanManualInputController.php
@@ -13,6 +13,7 @@ use Carbon\Carbon;
 use Illuminate\Http\Request;
 use App\Models\Inventory\UnitOfMeasure;
 use App\Enums\ProcurementPlanStatusEnum;
+use App\Policies\Procurement\PlanManualInputPolicy;
 
 
 class PlanManualInputController extends Controller
@@ -20,8 +21,9 @@ class PlanManualInputController extends Controller
     //
     public function index(Request $request)
     {
+         $this->authorize('viewAny', PlanLineItems::class);
         $planId = $request->query('plan_id');//todo pass plan id from url
-        $plans = ConsolidatedProcurementPlan::all();
+        $plans = ConsolidatedProcurementPlan::where('Status', ProcurementPlanStatusEnum::Draft)->get();
 
         $lineItemsQuery = PlanLineItems::with(['item', 'item.category','item.itemuom']);
 
@@ -41,6 +43,7 @@ class PlanManualInputController extends Controller
 
     public function create()
     {
+        $this->authorize('create', PlanLineItems::class);
         $plans = ConsolidatedProcurementPlan::where('Status', ProcurementPlanStatusEnum::Draft)->get();
         $items = Item::with('category','itemuom')->get();
         $budgetLines = BudgetMaster::all();
@@ -50,11 +53,11 @@ class PlanManualInputController extends Controller
 
     public function store(PlanManualInputRequest $request)
     {
+        $this->authorize('store', PlanLineItems::class);
 
         $validated = $request->validated();
         $user = $request->user();
 
-        // Fetch the related models (optional, can skip if you just want to save foreign keys)
         $item = $request->getItem();
         $category = $request->getCategory();//todo
 
@@ -90,6 +93,7 @@ class PlanManualInputController extends Controller
     public function edit($lineItemId)
     {
         $lineItem = PlanLineItems::with(['item', 'item.category'])->findOrFail($lineItemId);
+        $this->authorize('edit', $lineItem);
         $plans = ConsolidatedProcurementPlan::all();
         $items = Item::all();
         $categories = ItemCategory::all();
@@ -101,9 +105,10 @@ class PlanManualInputController extends Controller
     public function update(PlanManualInputRequest $request, $lineItemId)
     {
         $validated = $request->validated();
+        $user = $request->user();
 
         $lineItem = PlanLineItems::findOrFail($lineItemId);
-
+        $this->authorize('update', $lineItem);
         $lineItem->PlanID = $validated['PlanID'];
         $lineItem->ItemID = $validated['ItemID'];
         $lineItem->CategoryID = $validated['CategoryID'];
@@ -124,9 +129,11 @@ class PlanManualInputController extends Controller
         return redirect()->route('procurement.procurementplan.planconsolidation.manualentry.index')->with('success', 'Line item updated successfully.');
     }
 
-    public function destroy($lineItemId)
+    public function destroy(Request $request, $lineItemId)
     {
+        $user = $request->user();
         $lineItem = PlanLineItems::findOrFail($lineItemId);
+        $this->authorize('delete', $lineItem);
         $lineItem->delete();
         activity()->causedBy($user)->performedOn($lineItem)->event('deleted')->log('deleted plan line item ' . $lineItem->ItemID);
